정산 내역 s3 업로드 및 feign 클라이언트 적용 (#36)
* Feat: 결제 승인 구현

# Conflicts:
#	build.gradle
#	src/main/java/com/fifteen/auction/domain/payment/controller/PaymentController.java
#	src/main/java/com/fifteen/auction/domain/payment/service/PaymentService.java
#	src/main/java/com/fifteen/auction/global/controller/GlobalExceptionHandler.java
#	src/main/java/com/fifteen/auction/global/dto/error/ErrorCode.java
#	src/main/java/com/fifteen/auction/global/dto/exception/PaymentFailException.java

* Feat: 주문 정보 불러오기 구현

# Conflicts:
#	src/main/java/com/fifteen/auction/domain/payment/service/PaymentService.java

* Feat: 주문 취소 구현

* Feat: 구매 확정 구현, 주문 생성 코드 수정

* Feat: 결제 취소 구현 및 결제 요청 승인 과정 보완

* Feat: 결제 상세 내역 조회 구현 및 결제 취소 보완

* Feat: 즉시 정산 기능 구현

* Fix: 결제 취소를 제외한 결제 관련 기능들 수정

* Refactor: 불필요한 코드 제거 및 코드 정렬

* refactor(Settlement, Order): n+1 문제 해결 및 책임 분리

* refactor(Payment): Toss와 통신 관련 기능 분리 및 결제 관련 책임 분리

* fix(Payment, Settlement, Order): 에러 코드 수정

* Rebase dev

* refactor(Payment): feign 클라이언트 적용, 결제 구조 수정

* feat(Settlement): 정산 데이터 S3로 업로드 후 url로 반환 하는 기능 적용

* refactor(Settlement): Settlement 엔티티에 정산 내역 url필드 추가, charge 엔티티의 수수료 타입 변경

* refactor(Payment, Settlement, Order): 로깅 구현

* chore(Payment): 토스 시크릿키 추가

* Fix(Payment, Order, Settlement): 리베이스 후 오류 수정

* Fix(Payment, Order, Settlement): 추가 수정

* refactor(Payment, Order, Settlement): 파일 이름 수정 및 자동 정산 스케줄러 적용

* chore(Payment): 시크릿키 설정 변경

+++ src/main/java/com/fifteen/auction/domain/settlement/entity/Charge.java
@@ -6,6 +6,8 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -16,9 +18,9 @@ public class Charge {
     @Enumerated(EnumType.STRING)
     private ChargeType id;
 
-    private double proportion;
+    private BigDecimal proportion;
 
-    public void updateProportion(double proportion) {
+    public void updateProportion(BigDecimal proportion) {
         this.proportion = proportion;
     }
 }
